rename specific services. add and refine docs
