+++ packages/cli/src/modules/insights/__tests__/insights-collection.service.test.ts
@@ -22,7 +22,7 @@ import { createTeamProject } from '@test-integration/db/projects';
 import { createWorkflow } from '@test-integration/db/workflows';
 import * as testDb from '@test-integration/test-db';
 
-import { InsightsCollectionService } from '../insights.collection.service';
+import { InsightsCollectionService } from '../insights-collection.service';
 import { InsightsConfig } from '../insights.config';
 
 // Initialize DB once for all tests
